feat: add utility functions for streak calculation, photo upload, data export, and notifications
feat: implement chart drawing functionality for weekly mood and mood frequency

style: create a comprehensive CSS file for theming, animations, and responsive design

chore: initialize Vite configuration for project setup
