+++ .gitignore
@@ -1,0 +1,5 @@
+﻿node_modules/
+dist/
+.env
+*.log
+.DS_Store
