+++ .prettierrc.json
@@ -1,0 +1,5 @@
+﻿{
+  "semi": true,
+  "singleQuote": true,
+  "tabWidth": 2
+}
