+++ .vite/deps_temp_63222340/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "module"
+}
