+++ package.json
@@ -1,0 +1,13 @@
+{
+  "name": "pastel-daily-diary",
+  "version": "1.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^5.x.x"
+  }
+}
