+++ src/modules/diary.js
